refactor: remove ConversationLoggingPolicy and consolidate into core infrastructure
- Remove ConversationLoggingPolicy class and tests
- Refactor ToolCallBufferPolicy to be self-contained (no longer inherits from ConversationLoggingPolicy)
- Remove redundant logging - core hooks infrastructure already captures original→final versions
- Update CHANGELOG
