+++ src/luthien_proxy/policies/tool_call_buffer.py
@@ -4,26 +4,42 @@ from __future__ import annotations
 
 from copy import deepcopy
 from dataclasses import dataclass, field
-from typing import Any, AsyncIterator, Mapping
+from datetime import datetime, timezone
+from typing import Any, AsyncIterator, Mapping, Optional
 
 from luthien_proxy.types import JSONObject
 
-from .conversation_logger import ConversationLoggingPolicy, ConversationLogStreamContext, ToolCallState
+from .base import LuthienPolicy, StreamPolicyContext
 
 TOOL_CALL_DEBUG_TYPE = "conversation:tool-call"
 TOOL_CALL_SCHEMA = "luthien.conversation.tool_call.v1"
 
 
 @dataclass
-class ToolCallBufferContext(ConversationLogStreamContext):
-    """Extend conversation stream context with buffering metadata."""
+class ToolCallState:
+    """Incrementally collected state for a streaming tool call."""
 
+    identifier: str
+    call_type: str
+    name: str = ""
+    arguments: str = ""
+
+
+@dataclass
+class ToolCallBufferContext(StreamPolicyContext):
+    """Stream context with tool call tracking and buffering."""
+
+    content_parts: list[str] = field(default_factory=list)
+    tool_calls: dict[str, ToolCallState] = field(default_factory=dict)
+    tool_call_indexes: dict[int, str] = field(default_factory=dict)
+    finish_reason: Optional[str] = None
+    role: Optional[str] = None
     buffered_chunks: list[dict[str, Any]] = field(default_factory=list)
     tool_call_active: bool = False
     logged_tool_ids: set[str] = field(default_factory=set)
 
 
-class ToolCallBufferPolicy(ConversationLoggingPolicy):
+class ToolCallBufferPolicy(LuthienPolicy):
     """Intercept streaming tool calls, log them, then replay original chunks."""
 
     def create_stream_context(self, stream_id: str, request_data: dict[str, Any]) -> ToolCallBufferContext:
@@ -54,7 +70,6 @@ class ToolCallBufferPolicy(ConversationLoggingPolicy):
                 flushed = await self._flush_tool_calls(context)
                 for buffered in flushed:
                     yield buffered
-            await self._emit_stream_summary(context)
 
     async def async_post_call_success_hook(
         self,
@@ -62,11 +77,9 @@ class ToolCallBufferPolicy(ConversationLoggingPolicy):
         user_api_key_dict: Mapping[str, Any] | None,
         response: Mapping[str, Any],
     ) -> Mapping[str, Any]:
-        """Log non-stream tool calls in addition to base conversation logs."""
-        result = await super().async_post_call_success_hook(data, user_api_key_dict, response)
-
+        """Log non-stream tool calls."""
         if bool(data.get("stream")):
-            return result
+            return response
 
         tool_calls = self._extract_message_tool_calls(response)
         if tool_calls:
@@ -79,7 +92,7 @@ class ToolCallBufferPolicy(ConversationLoggingPolicy):
             )
             await self._record_debug_event(TOOL_CALL_DEBUG_TYPE, record)
 
-        return result
+        return response
 
     # ------------------------------------------------------------------
     # Streaming helpers
@@ -340,6 +353,214 @@ class ToolCallBufferPolicy(ConversationLoggingPolicy):
                 function_arguments = function_payload.get("arguments")
                 if isinstance(function_arguments, str):
                     state.arguments = function_arguments
+
+    # ------------------------------------------------------------------
+    # Helper methods (extracted from removed ConversationLoggingPolicy)
+    # ------------------------------------------------------------------
+    def _capture_stream_chunk(
+        self,
+        context: ToolCallBufferContext,
+        chunk: Mapping[str, Any],
+    ) -> None:
+        """Extract and update context state from a streaming chunk."""
+        choices = chunk.get("choices")
+        if not isinstance(choices, list) or not choices:
+            return
+        first = choices[0]
+        if not isinstance(first, Mapping):
+            raise TypeError("stream chunk choice must be a mapping")
+        finish_reason = first.get("finish_reason")
+        if finish_reason is not None:
+            if not isinstance(finish_reason, str):
+                raise TypeError("finish_reason must be a string when present")
+            context.finish_reason = finish_reason
+        delta = first.get("delta")
+        if delta is None:
+            return
+        if not isinstance(delta, Mapping):
+            raise TypeError("stream delta must be a mapping")
+        role = delta.get("role")
+        if role is not None:
+            if not isinstance(role, str) or not role:
+                raise ValueError("stream delta role must be a non-empty string")
+            context.role = role
+        content = delta.get("content")
+        if content is not None:
+            if not isinstance(content, str):
+                raise TypeError("stream delta content must be a string when present")
+            context.content_parts.append(content)
+        tool_calls = delta.get("tool_calls")
+        if tool_calls is not None:
+            if not isinstance(tool_calls, list):
+                raise TypeError("stream delta tool_calls must be a list")
+            for index, raw in enumerate(tool_calls):
+                if not isinstance(raw, Mapping):
+                    raise TypeError("tool call delta entries must be mappings")
+                identifier = self._resolve_tool_call_identifier(context, raw, index)
+                state = context.tool_calls.get(identifier)
+                if state is None:
+                    call_type = raw.get("type")
+                    if not isinstance(call_type, str) or not call_type:
+                        call_type = "function"
+                    call_id = raw.get("id")
+                    state = ToolCallState(identifier=call_id or identifier, call_type=call_type)
+                    context.tool_calls[identifier] = state
+                function_delta = raw.get("function")
+                if function_delta is not None:
+                    if not isinstance(function_delta, Mapping):
+                        raise TypeError("tool call function delta must be a mapping")
+                    name = function_delta.get("name")
+                    if name is not None:
+                        if not isinstance(name, str) or not name:
+                            raise ValueError("tool function name must be non-empty string")
+                        state.name = name
+                    arguments = function_delta.get("arguments")
+                    if arguments is not None:
+                        if not isinstance(arguments, str):
+                            raise TypeError("tool function arguments must be a string when present")
+                        state.arguments += arguments
+
+    def _resolve_tool_call_identifier(
+        self,
+        context: ToolCallBufferContext,
+        payload: Mapping[str, Any],
+        fallback_index: int,
+    ) -> str:
+        """Resolve stable identifier for a tool call from streaming deltas."""
+        call_index = payload.get("index")
+        raw_identifier = payload.get("id")
+        if isinstance(call_index, int):
+            mapped_identifier = context.tool_call_indexes.get(call_index)
+            if isinstance(raw_identifier, str) and raw_identifier:
+                if mapped_identifier is not None and mapped_identifier != raw_identifier:
+                    state = context.tool_calls.pop(mapped_identifier, None)
+                    if state is not None:
+                        state.identifier = raw_identifier
+                        context.tool_calls[raw_identifier] = state
+                context.tool_call_indexes[call_index] = raw_identifier
+                return raw_identifier
+            if mapped_identifier is not None:
+                return mapped_identifier
+
+        if isinstance(raw_identifier, str) and raw_identifier:
+            return raw_identifier
+
+        identifier = self._tool_call_identifier(payload, fallback_index)
+        if isinstance(call_index, int):
+            context.tool_call_indexes.setdefault(call_index, identifier)
+        return identifier
+
+    def _tool_call_identifier(self, payload: Mapping[str, Any], index: int) -> str:
+        """Generate fallback identifier for a tool call."""
+        identifier = payload.get("id")
+        if isinstance(identifier, str) and identifier:
+            return identifier
+        idx = payload.get("index")
+        if isinstance(idx, int):
+            return f"tool_{idx}"
+        return f"tool_{index}"
+
+    def _parse_tool_calls(self, value: Any, *, allow_empty: bool = False) -> list[dict[str, str]]:
+        """Parse tool_calls array from response message."""
+        if value is None:
+            return [] if allow_empty else self._unexpected_tool_call("value is None")
+        if not isinstance(value, list):
+            raise TypeError("tool_calls must be a list")
+        parsed: list[dict[str, str]] = []
+        for index, raw in enumerate(value):
+            if not isinstance(raw, Mapping):
+                raise TypeError("tool call entries must be mappings")
+            call_type = raw.get("type")
+            if not isinstance(call_type, str) or not call_type:
+                call_type = "function"
+            call_id = raw.get("id")
+            if call_id is not None and not isinstance(call_id, str):
+                raise TypeError("tool call id must be a string when present")
+            function = raw.get("function")
+            if not isinstance(function, Mapping):
+                raise ValueError("tool call missing function payload")
+            name = function.get("name")
+            if not isinstance(name, str) or not name:
+                raise ValueError("tool function name must be non-empty string")
+            arguments = function.get("arguments")
+            if not isinstance(arguments, str):
+                raise TypeError("tool function arguments must be a string")
+            parsed.append(
+                {
+                    "id": call_id or f"tool_{index}",
+                    "type": call_type,
+                    "name": name,
+                    "arguments": arguments,
+                }
+            )
+        return parsed
+
+    def _parse_legacy_function_call(self, payload: Any) -> dict[str, str]:
+        """Parse legacy function_call payload."""
+        if not isinstance(payload, Mapping):
+            raise TypeError("function_call payload must be a mapping")
+        name = payload.get("name")
+        if not isinstance(name, str) or not name:
+            raise ValueError("function_call name must be a non-empty string")
+        arguments = payload.get("arguments")
+        if not isinstance(arguments, str):
+            raise TypeError("function_call arguments must be a string")
+        raw_id = payload.get("id")
+        identifier = raw_id if isinstance(raw_id, str) and raw_id else "legacy_function_call"
+        return {
+            "id": identifier,
+            "type": "function",
+            "name": name,
+            "arguments": arguments,
+        }
+
+    def _require_call_id(self, payload: Mapping[str, Any]) -> str:
+        """Extract call_id from payload, raising error if not found."""
+        candidates = (
+            payload.get("litellm_call_id"),
+            self._metadata_lookup(payload, "litellm_call_id"),
+            self._params_lookup(payload, "litellm_call_id"),
+        )
+        for candidate in candidates:
+            if isinstance(candidate, str) and candidate:
+                return candidate
+        raise ValueError("payload missing litellm_call_id")
+
+    def _extract_trace_id(self, payload: Mapping[str, Any]) -> Optional[str]:
+        """Extract trace_id from payload."""
+        candidates = (
+            payload.get("litellm_trace_id"),
+            self._metadata_lookup(payload, "litellm_trace_id"),
+            self._params_lookup(payload, "litellm_trace_id"),
+        )
+        for candidate in candidates:
+            if isinstance(candidate, str) and candidate:
+                return candidate
+        return None
+
+    def _metadata_lookup(self, payload: Mapping[str, Any], key: str) -> Any:
+        """Look up value in metadata dict."""
+        metadata = payload.get("metadata")
+        if isinstance(metadata, Mapping):
+            return metadata.get(key)
+        return None
+
+    def _params_lookup(self, payload: Mapping[str, Any], key: str) -> Any:
+        """Look up value in litellm_params.metadata dict."""
+        params = payload.get("litellm_params")
+        if isinstance(params, Mapping):
+            metadata = params.get("metadata")
+            if isinstance(metadata, Mapping):
+                return metadata.get(key)
+        return None
+
+    def _timestamp(self) -> str:
+        """Return current UTC timestamp as ISO string."""
+        return datetime.now(timezone.utc).isoformat()
+
+    def _unexpected_tool_call(self, message: str) -> list[dict[str, str]]:
+        """Raise error for unexpected tool call payload."""
+        raise ValueError(f"unexpected tool call payload: {message}")
 
 
 __all__ = ["ToolCallBufferPolicy"]
